+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the "al-codeactions" extension will be documented in this file.
 
+## 1.0.26
+
+- fix for multiple `Extract to Label`: If multiple labels have been selected and the global variable was at the bottom of the file, the `rename` was executed at a wrong position. Related to #149
+
 ## 1.0.25
 
 - Enhancement to Extract to label: If there are further occurences of the same string literal, then you're asked if you want to replace these as well. #149